feat(moonbeam-app): #24 - adds initial draft of military status subscription.
Pending implementation of the actual Paywall which will prevent users from accessing
the application unless they have the right verification status. The follow-up PR/commit
will be comprised of implementing all UI components, and subscribing to this status
update from the dashboard and/or from the root of the application, instead of from the
registration component.

MOONBM-24