feat(getmoonbeam-app): #42 - improvements to store and profile picture loading.
This improves the laod time for offers in store, by introducing a "background"
loading task in the AppDrawer that will load up to 100 online and nearby offers,
as well as all of the Fidelis offers. In addition, this makes use of the expo-image
dependency to vastly reduce the time that it takes to load various images throughout
the app.
All that is pending now is making sure that our chosen prime brands are displayed
first.

MOONBM-42