feat(getmoonbeam-app): #22 - solved the transactional-updates issues.
We introduce a new FUNDED status for transactions to keep track of which
transactions have been paid out to Moonbeam. The CREDITED status which we
previously were using, will only be used while cashing out. We also enhanced
the statuses that Olive sends us and mapped them to appropriate internal statuses
according to the new Olive updates.

MOONBM-22